Encoder zeroing fixes attempted but not complete

+++ src/main/java/frc/robot/subsystems/DriveBase.java
@@ -3,9 +3,8 @@ package frc.robot.subsystems;
 import frc.robot.RobotMap;
 import edu.wpi.first.wpilibj.command.Subsystem;
 import frc.robot.commands.TankDrive;
-// import edu.wpi.first.wpilibj.Victor;
-// import edu.wpi.first.wpilibj.Talon;
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
+import com.ctre.phoenix.motorcontrol.can.WPI_VictorSPX;
 import edu.wpi.first.wpilibj.SpeedControllerGroup;
 import edu.wpi.first.wpilibj.ADXRS450_Gyro;
 import edu.wpi.first.wpilibj.drive.DifferentialDrive;
@@ -15,10 +14,18 @@ public class DriveBase extends Subsystem {
 
     public WPI_TalonSRX leftMiddleMaster;
     public WPI_TalonSRX rightMiddleMaster;
+
+    //Toaster/comp bot
+    // public WPI_VictorSPX leftFrontFollower;
+    // public WPI_VictorSPX leftRearFollower;
+    // public WPI_VictorSPX rightFrontFollower;
+    // public WPI_VictorSPX rightRearFollower;
+
+    //Practice bot
     public WPI_TalonSRX leftFrontFollower;
     public WPI_TalonSRX leftRearFollower;
     public WPI_TalonSRX rightFrontFollower;
-    public WPI_TalonSRX rightRearFollower;        
+    public WPI_TalonSRX rightRearFollower;
 
     public ADXRS450_Gyro gyro;
     private SpeedControllerGroup leftSide;
@@ -30,6 +37,14 @@ public class DriveBase extends Subsystem {
         super("DriveBase");
         this.leftMiddleMaster = new WPI_TalonSRX(RobotMap.leftMiddleMaster); 
         this.rightMiddleMaster = new WPI_TalonSRX(RobotMap.rightMiddleMaster);
+
+        //Toaster/comp bot
+        // this.leftFrontFollower = new WPI_VictorSPX(RobotMap.leftFrontFollower);
+        // this.leftRearFollower = new WPI_VictorSPX(RobotMap.leftRearFollower);
+        // this.rightFrontFollower = new WPI_VictorSPX(RobotMap.rightFrontFollower);
+        // this.rightRearFollower = new WPI_VictorSPX(RobotMap.rightRearFollower);
+
+        //Practice bot
         this.leftFrontFollower = new WPI_TalonSRX(RobotMap.leftFrontFollower);
         this.leftRearFollower = new WPI_TalonSRX(RobotMap.leftRearFollower);
         this.rightFrontFollower = new WPI_TalonSRX(RobotMap.rightFrontFollower);
@@ -57,11 +72,11 @@ public class DriveBase extends Subsystem {
         rightSide.set(0);
     }
 
-    public double getLeftPosition() {
+    public int getLeftPosition() {
         return leftMiddleMaster.getSelectedSensorPosition(0);
     }
 
-    public double getRightPosition() {
+    public int getRightPosition() {
         return rightMiddleMaster.getSelectedSensorPosition(0);
     }
 
@@ -72,6 +87,33 @@ public class DriveBase extends Subsystem {
     public void zeroEncoderPosition() {
         leftMiddleMaster.setSelectedSensorPosition(0,0,10);
         rightMiddleMaster.setSelectedSensorPosition(0,0,10);
+        //The below generally doesn't work.
+        // int leftPos = getLeftPosition();
+        // if (leftMiddleMaster.setSelectedSensorPosition(0,0,10).value != 0) {
+        //     System.out.println("ahhhhhh on left");
+        // }
+        // for(int i = 0; Math.abs(getLeftPosition()) > 100 && i < 1000; i++) {
+        //     if (i == 999) {
+        //         System.out.println("Took too long to zero left encoder (" + getLeftPosition() + ")");
+        //     } else {
+        //         if (leftPos != getLeftPosition()) {
+        //             System.out.println("LChanged from " + leftPos + " to " + getLeftPosition() + " at " + i);
+        //         }
+        //     }
+        // }
+        // int rightPos = getRightPosition();
+        // if (rightMiddleMaster.setSelectedSensorPosition(0,0,10).value != 0) {
+        //     System.out.println("ahjdsgfhg on right");
+        // }
+        // for(int i = 0; Math.abs(getRightPosition()) > 100 && i < 1000; i++) {
+        //     if (i == 999) {
+        //         System.out.println("Took too long to zero right encoder (" + getRightPosition() + ")");
+        //     } else {
+        //         if (rightPos != getRightPosition()) {
+        //             System.out.println("RChanged from " + rightPos + " to " + getRightPosition() + " at " + i);
+        //         }
+        //     }
+        // }
     }
 
     public void zeroGyroAngle() {
